+++ phpinsights.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 use NunoMaduro\PhpInsights\Domain\Insights\ForbiddenDefineGlobalConstants;
 use NunoMaduro\PhpInsights\Domain\Sniffs\ForbiddenSetterSniff;
+use PHP_CodeSniffer\Standards\Generic\Sniffs\CodeAnalysis\EmptyStatementSniff;
 use PHP_CodeSniffer\Standards\Generic\Sniffs\Files\LineLengthSniff;
 use PHP_CodeSniffer\Standards\Generic\Sniffs\PHP\NoSilencedErrorsSniff;
 use SlevomatCodingStandard\Sniffs\Functions\UnusedParameterSniff;
 use SlevomatCodingStandard\Sniffs\TypeHints\DisallowArrayTypeHintSyntaxSniff;
 use SlevomatCodingStandard\Sniffs\TypeHints\DisallowMixedTypeHintSniff;
-use SlevomatCodingStandard\Sniffs\TypeHints\PropertyTypeHintSniff;
 
 return [
 
@@ -78,7 +78,7 @@ return [
         NoSilencedErrorsSniff::class => [
             'exclude' => [
                 'src/Domain/Analyser.php',
-                'src/Domain/File.php',
+                'src/Domain/FileProcessors/FixerFileProcessor.php',
             ],
         ],
         ForbiddenDefineGlobalConstants::class => [
@@ -92,8 +92,27 @@ return [
                 'src/Domain/LinkFormatter/NullFileLinkFormatter.php',
             ],
         ],
-        PropertyTypeHintSniff::class => [
-            'enableNativeTypeHint' => false,
+        EmptyStatementSniff::class => [
+            'exclude' => [
+                'src/Domain/FileProcessors/FixerFileProcessor.php',
+            ],
         ],
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Requirements
+    |--------------------------------------------------------------------------
+    |
+    | Here you may define a level you want to reach per `Insights` category.
+    | When a score is lower than the minimum level defined, then an error
+    | code will be returned. This is optional and individually defined.
+    |
+    */
+
+    'requirements' => [
+        'min-quality' => 90.0,
+        'min-architecture' => 80.0,
+        'min-style' => 97.8,
     ],
 ];