+++ src/Application/Console/Definitions/AnalyseDefinition.php
@@ -53,9 +53,9 @@ final class AnalyseDefinition extends BaseDefinition
             new InputOption(
               'format',
               null,
-              InputOption::VALUE_REQUIRED,
+              InputOption::VALUE_IS_ARRAY | InputOption::VALUE_REQUIRED,
                 'Format to output the result in [console, json, checkstyle]',
-                'console'
+                ['console']
             ),
             new InputOption(
                 'fix',