+++ src/Application/Console/Formatters/Json.php
@@ -30,8 +30,6 @@ final class Json implements Formatter
     /**
      * Format the result to the desired format.
      *
-     * @param InsightCollection $insightCollection
-     * @param string $dir
      * @param array<string> $metrics
      *
      * @throws Exception
@@ -66,7 +64,6 @@ final class Json implements Formatter
     /**
      * Outputs the issues errors according to the format.
      *
-     * @param InsightCollection $insightCollection
      * @param array<string> $metrics
      *
      * @return array<string, array<int, array<string, int|string>>|null>
@@ -114,6 +111,7 @@ final class Json implements Formatter
                         'line' => $detail->hasLine() ? $detail->getLine() : null,
                         'function' => $detail->hasFunction() ? $detail->getFunction() : null,
                         'message' => $detail->hasMessage() ? $detail->getMessage() : null,
+                        'diff' => $detail->hasDiff() ? $detail->getDiff() : null,
                     ]);
                 }
             }