+++ src/Application/DirectoryResolver.php
@@ -16,7 +16,7 @@ final class DirectoryResolver
         /** @var string $directory */
         $directory = $input->getArgument('directory') ?? (string) getcwd();
 
-        if ($directory[0] !== DIRECTORY_SEPARATOR && preg_match('~\A[A-Z]:(?![^/\\\\])~i', $directory) === 0) {
+        if (isset($directory[0]) && $directory[0] !== DIRECTORY_SEPARATOR && preg_match('~\A[A-Z]:(?![^/\\\\])~i', $directory) === 0) {
             $directory = (string) getcwd() . DIRECTORY_SEPARATOR . $directory;
         }