+++ src/Domain/File.php
@@ -71,7 +71,7 @@ final class File extends BaseFile
             foreach ($this->tokenListeners[$token['code']] as $sniff) {
                 $this->activeSniff = $sniff;
                 try {
-                    @$sniff->process($this, $stackPtr);
+                    $sniff->process($this, $stackPtr);
                 } catch (\Throwable $e) {
                     $this->addError('Unparsable php code: syntax error or wrong phpdocs.', $stackPtr, $token['code']);
                 }