+++ src/Domain/Insights/SniffDecorator.php
@@ -72,6 +72,10 @@ final class SniffDecorator implements Sniff, Insight, HasDetails, Fixable
      */
     public function process(File $file, $stackPtr)
     {
+        set_error_handler(static function (): bool {
+            throw new \RuntimeException();
+        }, E_NOTICE);
+
         if ($file instanceof InsightFile && $this->skipFilesFromIgnoreFiles($file)) {
             return;
         }