+++ tests/TestCase.php
@@ -19,6 +19,7 @@ use ReflectionClass;
 use ReflectionException;
 use Symfony\Component\Console\Output\NullOutput;
 use Tests\Fakes\FakeFileRepository;
+use Tests\Fakes\FakeInput;
 
 abstract class TestCase extends BaseTestCase
 {
@@ -78,7 +79,7 @@ abstract class TestCase extends BaseTestCase
         array $filePaths,
         string $dir = ''
     ) : InsightCollection {
-        $config = ConfigResolver::resolve($config, $dir);
+        $config = ConfigResolver::resolve($config, FakeInput::directory($dir));
 
         $container = Container::make();
         \assert($container instanceof \League\Container\Container);